Fix/api authorization protect server api (#429)
* clone the request body so it can be reused

* add to all server api

* move more functions out of api
