+++ src/app/api/allNewRoutingChat/route.ts
@@ -2,13 +2,9 @@
 
 import { type ChatBody } from '@/types/chat'
 import { routeModelRequest } from '~/utils/streamProcessing'
-import { type NextRequest, type NextResponse } from 'next/server'
 import { buildPrompt } from '~/app/utils/buildPromptUtils'
-import { OpenAIError } from '~/utils/server'
-import {
-  withAppRouterAuth,
-  type AuthenticatedRequest,
-} from '~/utils/appRouterAuth'
+import { type AuthenticatedRequest } from '~/utils/appRouterAuth'
+import { withCourseAccessFromRequest } from '~/app/api/authorization'
 
 export const runtime = 'nodejs'
 export const dynamic = 'force-dynamic'
@@ -54,4 +50,4 @@ async function handler(req: AuthenticatedRequest) {
   }
 }
 
-export const POST = withAppRouterAuth(handler)
+export const POST = withCourseAccessFromRequest('any')(handler)
