+++ src/utils/streamProcessing.ts
@@ -2,9 +2,9 @@ import { type CoreMessage } from 'ai'
 import { type NextApiRequest, type NextApiResponse } from 'next'
 import posthog from 'posthog-js'
 import { v4 as uuidv4 } from 'uuid'
-import { runBedrockChat } from '~/app/api/chat/bedrock/route'
-import { runGeminiChat } from '~/app/api/chat/gemini/route'
-import { runSambaNovaChat } from '~/app/api/chat/sambanova/route'
+import { runBedrockChat } from '~/app/utils/bedrock'
+import { runGeminiChat } from '~/utils/modelProviders/routes/gemini'
+import { runSambaNovaChat } from '~/app/utils/sambanova'
 import { runAnthropicChat } from '~/app/utils/anthropic'
 import { runOllamaChat } from '~/app/utils/ollama'
 import { runVLLM } from '~/app/utils/vllm'
